pyproject.toml
[project]
name = "kopichat-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.60.0",
    "groq>=1.0.0",
    "pyaudio>=0.2.14",
]
